fix: Scarb.toml readme & license file not resolved
When the contract is being reoslved and a readme or license file is
specified in `Scarb.toml`, the contract resolving logic does not move
the readme & licnese file, causing a compilation error during `scarb
build`.

This PR fixes that by extracing information of these file from
`Scarb.toml` and also including them as file being moved to the
extracted sub-folder so builds won't fail due to it being missing.

Closes #22.