chore: add docs for asdf case (#38)
Add documentation regarding cases where asdf is used for version control
of the scarb version. This should help users that use asdf navigate the
usage of verifier.

This is however a temporary solution and in the long term we should
support asdf either via integrating into asdf ourselves or at least have
a version agnostic binary.