chore: remove scarb as dependency (WIP)
As we move towards having a universal verifier, we aim to reduce
dependencies that associate with a specific version of cairo. Scarb is
one such dependency that in the first place does not support being
imported as a library.

We want to first remove Scarb as our dependency, keeping those
that are supported as a library (`scarb-metadata` & `scarb-ui`), and try
to utilize cairo deps to replace their functionality. This should make
the maintenance easier and also reduce bugs that are caused by this
misuse of Scarb as a library.