chore: remove dynamic compiler (WIP)
Dynamic compiler was initially introduced in order to support compiling
multiple version of cairo with multiple version of scarb, and abstrating
away the versions from the CLI.

As we move to have a universal verifier, this should be removed and we
should instead start have generic crates to handle aspects of the
verification process.

As of right now we also do not use this for its original purpose
anymore, so it should be relatively safe to remove this.