fix: attempt to fix mod not found issue via delays (#69)
Currently there's a likely race condition when the dependencies are
being resolved, where the final build on the resolved projects fails
although the resolved project is correct. This is likely due to the
dependencies file updates that might happen when Scarb build happens
which causes some dependency to go missing in the middle of the build.

A quick fix is just to add a artificial delay in order to hopefully wait
until the files are fully updated before the build runs. This is of
course non-deterministic and not ideal, but we are aiming to address
this in the future by not relying on the files to be updated to build.