Universal tool rewrite (#80)
* chore: support 2.6.5

* Introduce command line interface

Also wrap most options in newtype wrapper and parse them instead of
validating.  Right now the network selection is as Subcommand but that
should be implemented using ValueEnum.

* Make ProjectDir hold an absolute path

* Move network selection to arguments from the subcommand

Made api module use the type defined in the args instead of the custom
one.

* Introduce Api client

Seems much cleaner than concatenating string.  Additionally retry with
exponential back off, by the way of external dependency.

Still there are 2 possible clients, public and private and not all
endpoints are available at both, but this is going to ba tackled in
the future.

* My notes

* Split into submission and status checking command

* Add license handling based on SPDX identifiers

* Store Scarb project metadata instead of project path

This lets us use this later without worrying and error checking.

* [wip]

* save progress

* complete rewrite

* Remove unused stuff

GitHub workflows that were placeholders, config files for tools that
weren't used.

* Get rid of workspace for such simple project

* more tweaks

* Documentation updates to the new verison

* Formatting

* Quickstart guide

* Perform dry run by default, need explicit flag to execute

Also sneak in the license name mapping for our current backend.

* Testnet -> Sepolia

* Document matching name requirement

* Formatting

* Comment on version support

---------

Co-authored-by: chen <23054115+cwkang1998@users.noreply.github.com>