Updates contract verifier to version 0.2.5 (#97)
Updates the starknet-contract-verifier version from 0.2.4 to 0.2.5.

Removes the explicit `DryRun` error and modifies the `submit` function to return a string instead of an error when the `--execute` flag is not provided, which constitutes a dry run.

Additionally, excludes test, example, and benchmark directories from the source package to reduce unnecessary data.