Adds support for custom API endpoints
Allows users to specify a custom API endpoint URL for interacting with the Voyager verification service.

This change replaces the reliance on predefined network configurations (mainnet, sepolia) with the flexibility to use a custom URL. This caters to users who may be using a private or beta version of the API, or require connecting to a specific instance.

The update introduces the `--url` argument for both `verify` and `status` subcommands, and makes `--network` optional when `--url` is provided.

Updates examples and documentation to reflect the changes.